+++ src/commands/memes/pov.ts
@@ -1,8 +1,8 @@
 import { CommandDefinition } from '../../lib/command';
-import { CommandCategory } from '../../constants';
+import { CommandCategory, imageBaseUrl } from '../../constants';
 import { makeEmbed } from '../../lib/embed';
 
-const POV_URL = `${process.env.IMAGE_BASE_URL}memes/pov_downscaled.gif`;
+const POV_URL = `${imageBaseUrl}/memes/pov_downscaled.gif`;
 
 export const pov: CommandDefinition = {
     name: 'pov',